requirements.txt
streamlit
streamlit-audiorecorder
pydub
torch
timm
tensorflow
tensorflow-hub
tf-keras
librosa
numpy
pathlib